apply refactoring to package manager utils
