+++ src/cli.ts
@@ -5,10 +5,12 @@ import { z } from 'zod';
 import { argumentParser } from 'zodcli';
 import type { ChalkInstance } from 'chalk';
 import chalk from 'chalk';
-import { getCurrentPackageManager } from './buildApplication/getCurrentPackageManager';
-import { getSpawnCommand, nextOnPagesVersion } from './utils';
-import type { PackageManager } from './utils';
-import { isWindows } from './utils/isWindows';
+import { nextOnPagesVersion } from './utils';
+import type { PackageManager } from './buildApplication/packageManagerUtils';
+import {
+	getCurrentPackageManager,
+	getPackageManagerSpawnCommand,
+} from './buildApplication/packageManagerUtils';
 
 // A helper type to handle command line flags. Defaults to false
 const flag = z
@@ -220,7 +222,7 @@ export async function printEnvInfo(): Promise<void> {
 
 function getPackageVersion(packageName: string): string {
 	try {
-		const command = isWindows() ? 'npm.cmd' : 'npm';
+		const command = getPackageManagerSpawnCommand('npm');
 		const commandOutput = execFileSync(
 			command,
 			['list', packageName, '--json', '--depth=0'],
@@ -238,7 +240,7 @@ function getPackageVersion(packageName: string): string {
 function getBinaryVersion(binaryName: PackageManager): string {
 	const commandArgs = ['--version'];
 	try {
-		return execFileSync(getSpawnCommand(binaryName), commandArgs)
+		return execFileSync(getPackageManagerSpawnCommand(binaryName), commandArgs)
 			.toString()
 			.trim();
 	} catch {
