+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './fs';
 export * from './version';
-export * from './getSpawnCommand';
 export * from './routing';
