--- src/utils/isWindows.ts
@@ -1,3 +1,0 @@
-export function isWindows(): boolean {
-	return process.platform === 'win32';
-}
